backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# Scheduling
apscheduler==3.10.4

# Free alternatives (no Google Cloud billing needed)
gtts==2.5.1                  # Text-to-Speech (free, no key)
deep-translator==1.11.4      # Translation (free, no key)
langdetect==1.0.9            # Language detection (free, local)
pytesseract==0.3.10          # OCR via Tesseract (free, local)
# Groq Whisper handles Speech-to-Text (uses existing GROQ_API_KEY)

# Twilio WhatsApp
twilio==9.0.5

# LLM (Groq / OpenAI / Gemini) – only need one
groq==0.9.0
openai==1.30.1
google-generativeai==0.6.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
pypdf2==3.0.1
python-jose==3.3.0
